### 2024-02-25 #234 @editorialbot
[PRE REVIEW]: The University of Toronto Climate Downscaling Workflow: Tools and Resources for Climate Change Impact Analysis
**Submitting author:** @mikemorris12 (<a href="http://orcid.org/0000-0002-5758-2182">Michael Morris</a>)
**Repository:** https://github.com/mikemorris12/UTCDW_Guidebook/
**Branch with paper.md** (empty if default branch): JOSE-submission
**Version:** v1.0
