### 2024-02-26 #234 @mikemorris12
> @mikemorris12 👋 Let me start with an apology: I am the editor-in-chief and I have neglected JOSE for several months due to plain overwhelm. I have cut down on other service and am back here with intention of reviving our little alt-journal. Thank you for your submission! Given the time that has passed, do you have updates to the paper?

Hi @labarba, thanks for considering our submission. There are indeed a few minor changes to the UTCDW-Guidebook repo that have been made since initial submission. The changes are only minor typo fixes, nothing particularly substantive regarding the content and no changes to paper.md. I've just now merged the changes from the main branch into the JOSE-Submission branch, I hope that's ok!
