### 2024-03-04 #234 @acocac
Hi @mikemorris12, thanks for submitting. I went through the resource and draft of the paper. I understand the novelty of the resource and endorse the use of scalable scientific python ecosystem e..g Pangeo.

Before passing this to reviewers, I suggest some improvements in the submitted paper according to [the guidelines for authors](https://openjournals.readthedocs.io/en/jose/submitting.html):

> Wordcount for `paper.md` is 1254
- The expected length is around 1000 words max. Please consider to rephrase or remove some sentences from the article.

- The paper would benefit to add a short text with the instructional design and the experience of use in teaching. The latter is briefly mentioned in the Acknowledge section.

Please feel free to consider above suggestions, in particular to adhere to the suggested length for JOSE papers.
