### 2024-03-04 #234 @acocac
> Hi @acocac, looking forward to hearing your thoughts on our work. My coauthors and I would like to suggest the following people as potential reviewers:
>
> * Brian Rose, from the Department of Atmospheric & Environmental Sciences at SUNY Albany.
> * Pascal Bourgault, Emilie Bresson, or Travis Logan from Ouranos.
> * Kevin Schwarzchild, from the Department of Earth and Environmental Sciences at Columbia University.
> * James Rising, from the School of Marine Science & Policy at the University of Delaware.
>
> Thanks again for considering our submission to JOSE.

Re the suggested reviewers, I've only found GitHub handles for Brian Rose, Pascal Bourgault and James Rising. According to your approval, I'll contact them after you address my initial observations of the submission.
