### 2024-03-20 dd6703c
fixed broken links in guidebook, added windows yml file, plus edits to JOSE paper
