+++ JOSE/paper.md
@@ -15,30 +15,34 @@ affiliations:
   index: 1
 - name: Department of Physical and Environmental Sciences, University of Toronto Scarborough
   index: 2
-date: October 2023
+date: March 2024
 bibliography: paper.bib
 ---
 
 # Summary
 
-The University of Toronto Climate Downscaling Workflow (UTCDW) is a resource designed to guide users through the process of producing their own statistically downscaled climate projections for the purpose of climate change impact analysis. The target audience includes graduate students or practitioners in engineering, the physical, mathematical, and computational sciences. Such individuals may work outside the field of climate science, but are interested in how their subject of study is sensitive to climate change. The main component of the UTCDW is a [Jupyter Book](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/README.html) we call the "UTCDW Guidebook". The Guidebook introduces users to basic climate science concepts and beginning-to-intermediate concepts in the application of statistical climate downscaling. It also works through the decisions that must be made when designing a climate change impact study. Finally, it demonstrates how to download climate data, do exploratory analysis and model validation, and provides examples of end-to-end workflows for small climate downscaling projects. The Guidebook was written for self-study by members of the target audience, though it may be included as a part of a course on climate change impact analysis for graduate students or upper-year undergraduate students in the applied sciences. All of the Guidebook source materials (Jupyter Notebooks, markdown files, and Python scripts) are accessible in the [UTCDW_Guidebook GitHub repository](https://github.com/mikemorris12/UTCDW_Guidebook).
+The University of Toronto Climate Downscaling Workflow (UTCDW) is a resource designed to teach users how to produce their own statistically downscaled climate projections for the purpose of climate change impact analysis. The target audience includes graduate students and practitioners in engineering, the physical, mathematical, and computational sciences, who are interested in how their subject of study is sensitive to climate change. The main component of the UTCDW is a [Jupyter Book](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/README.html) called the "UTCDW Guidebook". The Guidebook introduces users to basic climate science concepts and beginning-to-intermediate concepts in the application of statistical climate downscaling. It also works through the decisions that must be made when designing a climate change impact study. Finally, it demonstrates how to download climate data, do exploratory analysis and model validation, and provides examples of end-to-end workflows for small climate downscaling projects. The Guidebook was written for self-study by members of the target audience, though it may be included as a part of a course on climate change impact analysis for graduate students or upper-year undergraduate students in the applied sciences. All of the Guidebook source materials are accessible in the [UTCDW_Guidebook GitHub repository](https://github.com/mikemorris12/UTCDW_Guidebook).
 
 # Statement of Need
 
[…]
