+++ JOSE/paper.bib
@@ -1,31 +1,3 @@
-@article{cannon2015,
-	title = {Bias {Correction} of {GCM} {Precipitation} by {Quantile} {Mapping}: {How} {Well} {Do} {Methods} {Preserve} {Changes} in {Quantiles} and {Extremes}?},
-	volume = {28},
-	url = {http://journals.ametsoc.org/doi/10.1175/JCLI-D-14-00754.1},
-	doi = {10.1175/JCLI-D-14-00754.1},
-	number = {17},
-	journal = {Journal of Climate},
-	author = {Cannon, Alex J. and Sobie, Stephen R. and Murdock, Trevor Q.},
-	month = sep,
-	year = 2015,
-	pages = {6938--6959},
-}
-
-@article{werner2016,
-	title = {Hydrologic extremes – an intercomparison of multiple gridded statistical downscaling methods},
-	volume = {20},
-	issn = {1607-7938},
-	url = {https://hess.copernicus.org/articles/20/1483/2016/},
-	doi = {10.5194/hess-20-1483-2016},
-	number = {4},
-	urldate = {2022-12-07},
-	journal = {Hydrol. Earth Syst. Sci.},
-	author = {Werner, Arelia T. and Cannon, Alex J.},
-	month = apr,
-	year = 2016,
-	pages = {1483--1508},
-}
-
 @book{kotamarthi2021,
 	address = {Cambridge},
 	title = {Downscaling {Techniques} for {High}-{Resolution} {Climate} {Projections}: {From} {Global} {Change} to {Local} {Impacts}},
@@ -60,15 +32,6 @@
 	author = {Maraun, Douglas and Widmann, Martin},
 	year = 2018,
 	doi = {10.1017/9781107588783},
-}
-
-@misc{pcic2021,
-	address = {University of Victoria},
-	title = {Statistically {Downscaled} {Climate} {Scenarios}},
-	url = {https://data.pacificclimate.org/portal/downscaled_cmip6/map/},
-	author = {Pacific Climate Impacts Consortium},
-	month = dec,
-	year = 2021,
 }
 
 @misc{logan2023,
