+++ UTCDW_env_Windows.yml
@@ -1,0 +1,368 @@
+name: UTCDW-env
+channels:
+  - conda-forge
+  - defaults
+  - claut
+dependencies:
+  - anyio=4.3.0=pyhd8ed1ab_0
+  - argon2-cffi=23.1.0=pyhd8ed1ab_0
+  - argon2-cffi-bindings=21.2.0=py39ha55989b_4
+  - arrow=1.3.0=pyhd8ed1ab_0
+  - asciitree=0.3.3=py_2
+  - asttokens=2.4.1=pyhd8ed1ab_0
+  - async-lru=2.0.4=pyhd8ed1ab_0
+  - attrs=23.2.0=pyh71513ae_0
+  - aws-c-auth=0.7.16=hec1de76_6
+  - aws-c-cal=0.6.10=hd481e46_1
+  - aws-c-common=0.9.13=hcfcfb64_0
+  - aws-c-compression=0.2.18=hd481e46_1
+  - aws-c-event-stream=0.4.2=h0f06f08_4
+  - aws-c-http=0.8.1=hdb5aac5_5
+  - aws-c-io=0.14.5=h08270f9_1
+  - aws-c-mqtt=0.10.2=hfea8755_4
+  - aws-c-s3=0.5.2=h4b2095a_0
+  - aws-c-sdkutils=0.1.15=hd481e46_1
+  - aws-checksums=0.1.18=hd481e46_1
+  - aws-crt-cpp=0.26.2=h8492d2a_7
+  - aws-sdk-cpp=1.11.267=h93f5800_1
+  - azure-core-cpp=1.11.1=h249a519_1
+  - azure-storage-blobs-cpp=12.10.0=h91493d7_1
+  - azure-storage-common-cpp=12.5.0=h91493d7_4
+  - babel=2.14.0=pyhd8ed1ab_0
+  - backcall=0.2.0=pyh9f0ad1d_0
+  - beautifulsoup4=4.12.3=pyha770c72_0
+  - bleach=6.1.0=pyhd8ed1ab_0
+  - blosc=1.21.5=hdccc3a2_0
+  - bokeh=3.3.4=pyhd8ed1ab_0
+  - boltons=23.1.1=pyhd8ed1ab_0
+  - bottleneck=1.3.8=py39hd88c2e4_0
+  - branca=0.7.1=pyhd8ed1ab_0
+  - brotli=1.1.0=hcfcfb64_1
+  - brotli-bin=1.1.0=hcfcfb64_1
+  - brotli-python=1.1.0=py39h99910a6_1
+  - bzip2=1.0.8=hcfcfb64_5
+  - c-ares=1.27.0=hcfcfb64_0
+  - ca-certificates=2024.2.2=h56e8100_0
+  - cached-property=1.5.2=hd8ed1ab_1
+  - cached_property=1.5.2=pyha770c72_1
+  - cairo=1.18.0=h1fef639_0
+  - cartopy=0.22.0=py39h32e6231_1
+  - certifi=2024.2.2=pyhd8ed1ab_0
+  - cf_xarray=0.9.0=pyhd8ed1ab_0
+  - cffi=1.16.0=py39ha55989b_0
+  - cfitsio=4.3.1=h9b0cee5_0
+  - cftime=1.6.3=py39hd88c2e4_0
+  - charset-normalizer=3.3.2=pyhd8ed1ab_0
+  - click=8.1.7=win_pyh7428d3b_0
+  - click-plugins=1.1.1=py_0
+  - cligj=0.7.2=pyhd8ed1ab_1
+  - cloudpickle=3.0.0=pyhd8ed1ab_0
+  - colorama=0.4.6=pyhd8ed1ab_0
[…]
