### 2024-03-20 #234 @mikemorris12
@acocac I have seen your messages, and I apologize for the delay. My coauthors and I recently hosted a hackathon based around the UTCDW Guidebook materials, and since you asked for discussion of experience using the materials for teaching, we felt it would be best to include a brief discussion of how the hackathon participants responded to the Guidebook in this new section of the paper. I am awaiting approval from the coauthors before submitting the revised paper, it shouldn't take much longer until we're ready to proceed.
