### 2024-04-16 #234 @acocac
Hi @brian-rose — the primary author of a submission to JOSE (The Journal of Open Source Education) has suggested you as a potential reviewer. Could you take a look at this submission and let us know if you might be able to volunteer a review for us? We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time.

Thanks! 🙏

Title: "The University of Toronto Climate Downscaling Workflow: Tools and Resources for Climate Change Impact Analysis" [View article proof on GitHub](https://raw.githubusercontent.com/openjournals/jose-papers/jose.00234/jose.00234/10.21105.jose.00234.pdf) 📄 👈
