### 2024-04-16 #234 @acocac
> Hi @acocac , I am able to review this paper, and willing to do so. I see that you have guidelines for reviewers and at first glance, I think I can do that!

Great - Thanks for the quick and positive response! JOSE reviews are checklist-driven to facilitate the assessment of the submission. We should wait the acceptance of another reviewer to start the review.
