### 2024-05-02 #234 @acocac
> @acocac I can do this review! Sorry I got buried in tasks and requests in the second half of my semester but starting to dig myself out now. I should be able to get this done next week.

Hi @brian-rose, thanks for the update. It's great to have you as a reviewer of this submission contributing to the climate science community!

I'll open a REVIEW issue to start the formal review.
