### 2024-05-02 #243 @editorialbot
[REVIEW]: The University of Toronto Climate Downscaling Workflow: Tools and Resources for Climate Change Impact Analysis
**Submitting author:** @mikemorris12 (<a href="http://orcid.org/0000-0002-5758-2182">Michael Morris</a>)
**Repository:** https://github.com/mikemorris12/UTCDW_Guidebook/
**Branch with paper.md** (empty if default branch): master
**Version:** v1.1.1
