### 2024-06-07 #243 @brian-rose
A couple of quick issues I noted in a first pass, before taking a deeper dive into the content:
1. The [repository](https://github.com/mikemorris12/UTCDW_Guidebook) does not have any releases, so it does not match the given v1.0 version number.
2. One of the references in the JOSE paper is to the documentation for the xclim package. It's a [valid link](https://xclim.readthedocs.io/en/stable/index.html) but not a DOI. The [suggestions within the xclim docs](https://xclim.readthedocs.io/en/stable/readme.html#how-to-cite-this-library) is to cite their JOSS paper (https://doi.org/10.21105/joss.05415). I think that would be a better choice.
