### 2024-06-14 #243 @aulemahal
I have finished going through the guidebook!

From the review checklist, the only remaining "issue" is indeed that there seems to be no versioning on the guidebook repo.
Otherwise, I find this document to be quite interesting. It is well written and complete. It is long, but (almost) no parts seemed superfluous. The topic is complex and to learn both the theory and the practical stuff, a user needs to take their time going though the book.

The following comments are not really publication-blocking to me.

The only section I found tedious was the data download. Station stuff is concise enough with  `ec3`  and the usage of `siphon` on PAVICS is clear too. However, all those custom functions with  `request` for the other dataset make the code look overly complex (too me). I would suggest the following ideas to make this part less cumbersome:
- Use `intake-esm` with Google's CMIP6 catalog. [See this doc](https://pangeo-data.github.io/pangeo-cmip6-cloud/accessing_data.html#loading-an-esm-collection)
- Redirect the readers to the HTML interfaces of ESGF and quickly show how to download the netCDFs by hand or by generated `wget` scripts (or even globus).

All my other comments are on details of the python code. Should I put them elsewhere ? As issues in the original repo ?

- I think there is too much usage of `pandas`. Xarray and Pandas have very similar but different APIs, as to not confuse the readers, I would suggest showing early in the document how to transform `DataFrames` into `Datasets` cleanly and only computing and plotting with xarray from then on.
- Similarly, `dask.array.apply_along_axis` is used in many places. I would suggest using `xr.apply_ufunc` instead (with `vectorize=True`). Also, the first place it is used (section 3.2), the data isn't even in a dask array.
- OpenDAP : It is mentionned that OpenDAP and PAVICS's THREDDS server are limited when it comes to large datasets. That's only true because no chunking is being given to `xr.open_dataset`. [PAVICS](https://pavics.ouranos.ca/datasets.html#b) suggests some chunking for its principal datasets and for others,  `chunks='auto'` might be ok for what this guidebook tries to show.
- For the observations, you could explore using the AHCCD dataset on PAVICS instead of downloading raw station data.
- However, this previous suggestion would require some quick explanation on how dask works and how to configure it. Showing to do `dask.config.set(num_workers=XX)` might save some headaches and machine jamming. A bonus section or links to tutorials about the distributed `Client` (and its dashboard!) might be useful.
- PCIC's download URL is down. Their website has been reorganized and the data is now accessible through an interactive portal (which resonates with my previous point).
- `xclim.ensembles.create_ensemble` can merge datasets along dimensions like `time`, the input arguments simply need to be given appropriately.
- Please use `xclim.atmos` instead of `xclim.indices`.
- Some cells have long content. Consider making multiline calls, as when reading the generated HTML, it is sometimes nicer if the user doesn't have to scroll left and right.
