### 2024-06-17 #243 @brian-rose
Small grammatical fix in the JOSE paper text: in the Guidebook Content section, line 51 of the rendered PDF. Replace

> Chapter 2 explains how climate change projections are made and what their limitations.

with

> Chapter 2 explains how climate change projections are made and what their limitations are.
