### 2024-06-17 mikemorris12/UTCDW_Guidebook#2 @brian-rose
Data access problem in notebook 3.2
As part of my [review of the JOSE submission](https://github.com/openjournals/jose-reviews/issues/243), I am running through all the notebooks. I found an error in [section 3.2.5.2 Downloading Complete Datasets](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter3/section3.2_gridded_obs.html#downloading-complete-datasets). The call to
```
requests.get("http://tools.pacificclimate.org/dataportal/gridded_observations/catalog/catalog.json")
```
produces
```
NameResolutionError: <urllib3.connection.HTTPConnection object at 0x307df8800>: Failed to resolve 'tools.pacificclimate.org' ([Errno 8] nodename nor servname provided, or not known)
```

Seems like `http://tools.pacificclimate.org/dataportal/gridded_observations/catalog/catalog.json` is not a valid URL ... perhaps the PCIC has made some changes to their public-facing API?
