### 2024-06-17 mikemorris12/UTCDW_Guidebook#3 @brian-rose
Unit errors in calls to xclim in Section 4
As part of my [review of the JOSE submission](https://github.com/openjournals/jose-reviews/issues/243) I'm going through all the notebooks.

The code is breaking in [Section 4.4.4 Extreme Indicators](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter4/section4.4_downscaling_validation.html#extreme-indicators).

The line
```
tn_obs_train = xci.tn_days_above(data_obs_train)
```
is failing with
```
AttributeError: 'NoneType' object has no attribute 'Unit'
```

I'm not familiar with xclim and I'm not too sure what's going wrong here. I see that @aulemahal suggested avoiding `xclim.indices` (imported as `xci` here), maybe this is related?

Full traceback:
```
Cell In[23], line 1
----> 1 tn_obs_train = xci.tn_days_above(data_obs_train)

File <boltons.funcutils.FunctionBuilder-97>:2, in tn_days_above(tasmin, thresh, freq, op)

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py:1293](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py#line=1292), in declare_units.<locals>.dec.<locals>.wrapper(*args, **kwargs)
   1289     check_units(bound_args.arguments.get(name), dim)
   1291 out = func(*args, **kwargs)
-> 1293 _check_output_has_units(out)
   1295 return out

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py:1105](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py#line=1104), in _check_output_has_units(out)
   1103 if "units" not in outd.attrs:
   1104     raise ValueError("No units were assigned in one of the indice's outputs.")
-> 1105 outd.attrs["units"] = ensure_cf_units(outd.attrs["units"])

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py:198](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py#line=197), in ensure_cf_units(ustr)
    193 def ensure_cf_units(ustr: str) -> str:
    194     """Ensure the passed unit string is CF-compliant.
    195
    196     The string will be parsed to pint then recast to a string by xclim's `pint2cfunits`.
    197     """
--> 198     return pint2cfunits(units2pint(ustr))

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py:190](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/xclim/core/units.py#line=189), in pint2cfunits(value)
    188 with warnings.catch_warnings():
    189     warnings.simplefilter("ignore", category=DeprecationWarning)
--> 190     return f"{value:cf}".replace("dimensionless", "")

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/pint/facets/plain/unit.py:67](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/pint/facets/plain/unit.py#line=66), in PlainUnit.__format__(self, spec)
     66 def __format__(self, spec: str) -> str:
---> 67     return self._REGISTRY.formatter.format_unit(self, spec)

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/pint/delegates/formatter/full.py:129](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/pint/delegates/formatter/full.py#line=128), in FullFormatter.format_unit(self, unit, uspec, sort_func, **babel_kwds)
    127 uspec = uspec or self.default_format
    128 sort_func = sort_func or self.default_sort_func
--> 129 return self.get_formatter(uspec).format_unit(
    130     unit, uspec, sort_func=sort_func, **babel_kwds
    131 )

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/pint/delegates/formatter/_to_register.py:102](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/pint/delegates/formatter/_to_register.py#line=101), in register_unit_format.<locals>.wrapper.<locals>.NewFormatter.format_unit(self, unit, uspec, **babel_kwds)
     99 else:
    100     units = self._registry.UnitsContainer(numerator)
--> 102 return func(units, registry=self._registry)

File [~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/cf_xarray/units.py:38](http://localhost:8888/~/miniconda3/envs/UTCDW-env/lib/python3.12/site-packages/cf_xarray/units.py#line=37), in short_formatter(unit, registry, **options)
     35 import re
     37 # convert UnitContainer back to Unit
---> 38 unit = registry.Unit(unit)
     39 # Print units using abbreviations (millimeter -> mm)
     40 s = f"{unit:~D}"

AttributeError: 'NoneType' object has no attribute 'Unit'
```
