### 2024-06-17 mikemorris12/UTCDW_Guidebook#3 @brian-rose
Update: the same error also occurs in
- [Section 4.5.1 Calculating Tropical Nights (Again)](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter4/section4.5_assessing_downscaled_projections.html#calculating-tropical-nights-again)
- [Section 6.1.3 Assessing Model-Obs Consistency](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter6/section6.1_station_1D_example_onesim.html#assessing-model-obs-consistency)
- [Section 6.2.2 Exploratory Multi-Model Analysis](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter6/section6.2_station_1D_example_multimodel.html#exploratory-multi-model-analysis)
- [Section 6.5.3.1 Model/Obs Consistency](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter6/section6.5_gridded_DBCCA_example.html#model-obs-consistency)
- [Section 6.4.2 Exploratory Analysis](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter6/section6.4_station_1D_example_nearterm.html#exploratory-analysis) after implementing the fix described in #4
