### 2024-06-17 mikemorris12/UTCDW_Guidebook#4 @brian-rose
Error in result of call to scipy.stats.ttest_ind_from_stats
As part of my [review of the JOSE submission](https://github.com/openjournals/jose-reviews/issues/243) I'm going through all the notebooks.

In [Section 6.4.2 Exploratory Analysis](https://utcdw.physics.utoronto.ca/UTCDW_Guidebook/Chapter6/section6.4_station_1D_example_nearterm.html#exploratory-analysis) I'm getting a failure after the call to `scipy.stats.ttest_ind_from_stats`. The code in the notebook is
```
# perform two_sample t-test to see if future temperatures are higher than past
pvals = []
for scenario in tas_ens_future_raw_mean.scenario.values:
    tstat, pval = stats.ttest_ind_from_stats(tas_ens_future_raw_mean.sel(scenario = scenario),
                                            tas_ens_future_raw_stdev.sel(scenario = scenario),
                                            neff_future_raw.sel(scenario = scenario),
                                            tas_ens_hist_raw_mean,
                                            tas_ens_hist_raw_stdev,
                                            neff_hist_raw,
                                            equal_var = False,
                                            alternative = 'greater')
    pvals.append(pval.values)

    print('p-values for mean change of raw tas')
```

Taking `pval.values` throws the error
```
AttributeError: 'numpy.float64' object has no attribute 'values'
```

This seems consistent with the [scipy docs](https://docs.scipy.org/doc/scipy/reference/generated/scipy.stats.ttest_ind_from_stats.html) which state that the return values `tstat`, `pval` should be float or array.
