### 2024-06-17 mikemorris12/UTCDW_Guidebook#4 @brian-rose
I changed the line
```
    pvals.append(pval.values)
```
to simply
```
    pvals.append(pval)
```
and the cell seemed to execute normally.
