### 2024-06-20 ba5041b
revisions from JOSE review feedback - issue building 6.6
