### 2024-06-21 #243 @acocac
> Thanks @brian-rose and @aulemahal for your helpful feedback! Apologies for all of the Python bugs, we've had quite a few users run the code without error so I wonder if there are package version issues that need to be sorted - I'll look into it. @acocac how do we proceed from here? I'm not used to this sort of open review process, so do I start making the revisions right away or will you write an editor's decision like a traditional journal? I'll wait to hear from you before I proceed.

Hi @mikemorris12, following the [review process](https://openjournals.readthedocs.io/en/jose/submitting.html#the-review-process) in the Submitting a paper to JOSE section, you should:

- Implement changes in the JOSE-submission branch.
- Respond to reviewer-raised issues (if any are raised) on the submission repository’s issue tracker. Reviewer and editor contributions, like any other contributions, should be acknowledged in the repository i.e. tag @[reviewer-github-handle] in commit messages.
- Upon successful completion of the review, merge the JOSE-submission branch into master and release a new version. I suggest Version 1.1.0 due to the minor nature of the comments (e.g. improve the data download using `intake-esm`). Please refer to [semantic versioning 2.0.0](https://semver.org/#summary) for further info.
- Remember to deposit a copy of your (updated) repository with a data-archiving service such as [Zenodo](https://zenodo.org/) or [figshare](https://figshare.com/), get a DOI for the archive, and update the review issue thread with the DOI.
- Tag me when you complete above steps. Then, I'll suggest the reviewers to revisit their checklists.
