### 2024-06-21 f8a9274
changed conda setup instructions and yml file per JOSE comment from @brian-rose
