+++ UTCDW_env.yml
@@ -1,0 +1,31 @@
+name: UTCDW-env2
+channels:
+  - conda-forge
+  - defaults
+  - claut
+dependencies:
+  - numpy
+  - scipy
+  - matplotlib
+  - seaborn
+  - cartopy
+  - nc-time-axis
+  - pandas
+  - jupyter
+  - xarray
+  - netCDF4
+  - zarr
+  - ec3
+  - esgf-pyclient
+  - gcsfs
+  - siphon
+  - requests
+  - tqdm
+  - dask
+  - dask-ml
+  - scikit-learn
+  - xesmf
+  - xclim
+  - xskillscore
+  - libpysal 
+  - esda
