+++ UTCDW_env_Linux.yml
@@ -1,393 +1,0 @@
-name: UTCDW-env
-channels:
-  - conda-forge
-  - defaults
-  - claut
-dependencies:
-  - _libgcc_mutex=0.1=conda_forge
-  - _openmp_mutex=4.5=2_gnu
-  - aiohttp=3.8.5=py39hd1e30aa_0
-  - aiosignal=1.3.1=pyhd8ed1ab_0
-  - alsa-lib=1.2.8=h166bdaf_0
-  - anyio=3.7.0=pyhd8ed1ab_1
-  - argon2-cffi=21.3.0=pyhd8ed1ab_0
-  - argon2-cffi-bindings=21.2.0=py39hb9d737c_3
-  - asciitree=0.3.3=py_2
-  - asttokens=2.2.1=pyhd8ed1ab_0
-  - async-timeout=4.0.2=pyhd8ed1ab_0
-  - attr=2.5.1=h166bdaf_1
-  - attrs=23.1.0=pyh71513ae_1
-  - aws-c-auth=0.6.30=h435f46f_0
-  - aws-c-cal=0.6.0=h93469e0_0
-  - aws-c-common=0.8.23=hd590300_0
-  - aws-c-compression=0.2.17=h862ab75_1
-  - aws-c-event-stream=0.3.1=h1e03375_0
-  - aws-c-http=0.7.10=h9ab9c9b_2
-  - aws-c-io=0.13.27=h3870b5a_0
-  - aws-c-mqtt=0.8.13=hd4f18eb_5
-  - aws-c-s3=0.3.12=he51b3f3_1
-  - aws-c-sdkutils=0.1.11=h862ab75_1
-  - aws-checksums=0.1.16=h862ab75_1
-  - aws-crt-cpp=0.20.2=h4389a34_17
-  - aws-sdk-cpp=1.10.57=h7b9373a_16
-  - backcall=0.2.0=pyh9f0ad1d_0
-  - backports=1.0=pyhd8ed1ab_3
-  - backports.functools_lru_cache=1.6.5=pyhd8ed1ab_0
-  - beautifulsoup4=4.12.2=pyha770c72_0
-  - bleach=6.0.0=pyhd8ed1ab_0
-  - blinker=1.6.2=pyhd8ed1ab_0
-  - blosc=1.21.4=h0f2a231_0
-  - bokeh=3.2.0=pyhd8ed1ab_0
-  - boltons=23.0.0=pyhd8ed1ab_0
-  - bottleneck=1.3.7=py39h389d5f1_0
-  - brotli=1.0.9=h166bdaf_9
-  - brotli-bin=1.0.9=h166bdaf_9
-  - brotli-python=1.0.9=py39h5a03fae_9
-  - brotlipy=0.7.0=py39hb9d737c_1005
-  - bzip2=1.0.8=h7f98852_4
-  - c-ares=1.19.1=hd590300_0
-  - ca-certificates=2023.7.22=hbcca054_0
-  - cachetools=4.2.4=pyhd8ed1ab_0
-  - cairo=1.16.0=hbbf8b49_1016
-  - cartopy=0.21.1=py39h4bd5d67_1
-  - cattrs=23.1.2=pyhd8ed1ab_0
-  - certifi=2023.7.22=pyhd8ed1ab_0
-  - cf_xarray=0.8.2=pyhd8ed1ab_0
-  - cffi=1.15.1=py39he91dace_3
-  - cftime=1.6.2=py39h2ae25f5_1
-  - charset-normalizer=3.1.0=pyhd8ed1ab_0
-  - click=8.1.3=unix_pyhd8ed1ab_2
-  - cloudpickle=2.2.1=pyhd8ed1ab_0
[…]
