### 2024-06-21 mikemorris12/UTCDW_Guidebook#4 @mikemorris12
Thanks for pointing out this bug. I didn't encounter it since when I ran it, the function `scipy.stats.ttest_ind_from_stats` returned the p-values in an `xr.DataArray`, from which I extracted the values in a `numpy` array using the `.values` accessor. Now it seems to be converting the output to a `numpy` array, so I've incorporated the suggested change in commit [ba5041b](https://github.com/mikemorris12/UTCDW_Guidebook/commit/ba5041b86f8eb3db55dcf53aedf11e5019dd0340). Thanks!
