### 2024-06-21 mikemorris12/UTCDW_Guidebook#3 @mikemorris12
Following the suggestion from @aulemahal I've changed the functions used to calculate Tropical Nights, Cold Spell Duration Index, and Cooling Degree Days to those from `xclim.indicators.atmos` instead of those from `xclim.indices`. This was done in commit [ba5041b](https://github.com/mikemorris12/UTCDW_Guidebook/commit/ba5041b86f8eb3db55dcf53aedf11e5019dd0340).

However, I was not able to reproduce the error even when using the `xclim.indices` function so I can't yet close the issue. Could you try running the new code and see if the error remains?
