### 2024-06-21 857e1fa
re-named 3.5 to reflect ESFG. Added 3.6 showing model data access with intake-esm, per @brian-rose and @aulemahal suggestion.
