### 2024-06-21 77a1071
re-built book following addition of section 6.3 (intake-esm)
