+++ _toc.yml
@@ -18,6 +18,7 @@ chapters:
     - file: Chapter3/section3.3_reanalysis
     - file: Chapter3/section3.4_climate_model_output_from_ESGF
     - file: Chapter3/section3.5_model_data_on_google_cloud
+    - file: Chapter3/section3.6_model_data_with_intake_esm
     - file: Chapter3/chapter3_summary
   - file: Chapter4/chapter4
     sections:
