### 2024-06-24 mikemorris12/UTCDW_Guidebook#3 @brian-rose
Hi @mikemorris12, I confirm that this is resolved! I ran the updated notebooks and things worked as expected. So I think you can close this issue.
