### 2024-06-24 mikemorris12/UTCDW_Guidebook#3 @brian-rose
Or rather, I will close the issue since I opened it in the first place!
