### 2024-06-25 mikemorris12/UTCDW_Guidebook#3 @aulemahal
Sorry I'm late, but the initial issue here is a breaking change in `pint` that broke the unit management of `cf_xarray` (that xclim uses). Pint 0.24.1 was released yesterday with a fix.
