### 2024-06-26 5d20da7
Fixed name of ch3.6. Added tutorial on accessing AHCCD obs to 3.1, per @ aulemahal suggestion in JOSE review.
