+++ UTCDW_env.yml
@@ -30,3 +30,4 @@ dependencies:
   - libpysal 
   - esda
   - intake-esm
+  - clisops
