### 2024-06-26 01d56e1
removed specific installation info from UTCDW_env_Windows.yml like I did for the other YML file.
