### 2024-06-27 229f9b3
updated typo and xclim reference in paper.md, per the feedback of @brian-rose and @aulemahal
