+++ JOSE/paper.md
@@ -31,7 +31,7 @@ Other educational materials on climate change impact assessment are available, b
 
 # Guidebook Content
 
-The UTCDW Guidebook consists of six chapters. The first four chapters introduce the reader the background knowledge required to perform downscaling. Chapter 1 serves as an introduction, and guides users in configuring their Python environment. Chapter 2 explains how climate change projections are made and what their limitations. Chapter 3 demonstrates how to access observational, reanalysis, and climate model data, and how to do exploratory analysis with observational and raw climate model data. Chapter 4 contains explanations of various methods of bias-correction and downscaling, and examples of how to validate historical downscaled model output and assess future projections. Python code for the downscaling methods is provided, leveraging the utility of ``xclim`` [@logan2023] and the [Pangeo](https://pangeo.io/packages.html) software ecosystem.
+The UTCDW Guidebook consists of six chapters. The first four chapters introduce the reader the background knowledge required to perform downscaling. Chapter 1 serves as an introduction, and guides users in configuring their Python environment. Chapter 2 explains how climate change projections are made and what their limitations are. Chapter 3 demonstrates how to access observational, reanalysis, and climate model data, and how to do exploratory analysis with observational and raw climate model data. Chapter 4 contains explanations of various methods of bias-correction and downscaling, and examples of how to validate historical downscaled model output and assess future projections. Python code for the downscaling methods is provided, leveraging the utility of ``xclim`` [@bourgault2023] and the [Pangeo](https://pangeo.io/packages.html) software ecosystem.
 
 The next two chapters focus on applying the content of Chapters 1-4. Chapter 5 introduces the "Downscaling Workflow" part of the UTCDW. This chapter unpacks the decisions that must be made when designing a climate change impact study and breaks the analysis tasks down into digestible steps. Chapter 6 contains examples of the workflow. Each example demonstrates the process of stating a problem, acquiring the necessary data, producing calibrated climate projections, and quantifying uncertainty. The examples focus on Canadian regions and observational data products, though we include links to available observational and downscaled data products for additional regions as well as global data. We welcome user submission of additional worked examples to contribute to the gallery of examples to be included on the website, or future versions of the UTCDW Guidebook
 
