+++ JOSE/paper.bib
@@ -56,3 +56,20 @@
 	year = 2016,
 	pages = {1937--1958},
 }
+
+@article{bourgault2023,
+	title = {xclim: xarray-based climate data analytics},
+	volume = {8},
+	issn = {2475-9066},
+	shorttitle = {xclim},
+	url = {https://joss.theoj.org/papers/10.21105/joss.05415},
+	doi = {10.21105/joss.05415},
+	language = {en},
+	number = {85},
+	urldate = {2023-11-09},
+	journal = {Journal of Open Source Software},
+	author = {Bourgault, Pascal and Huard, David and Smith, Trevor James and Logan, Travis and Aoun, Abel and Lavoie, Juliette and Dupuis, Eric and Rondeau-Genesse, Gabriel and Alegre, Raquel and Barnes, Clair and Laperriere, Alexis Beaupre and Biner, Sebastien and Caron, David and Ehbrecht, Carsten and Fyke, Jeremy and Keel, Tom and Labonte, Marie-Pier and Lierhammer, Ludwig and Low, Jwen-Fai and Quinn, Jamie and Roy, Philippe and Squire, Dougie and Stephens, Ag and Tanguy, Maliko and Whelan, Christopher},
+	month = may,
+	year = {2023},
+	pages = {5415},
+}
