### 2024-06-27 #243 @mikemorris12
Hi @acocac, I think I've addressed each of the reviewer’s comments. Here is a line-by-line response:

For @aulemahal ‘s comments:

-	Data download: Instead of replacing the `esgf-pyclient` and Google Cloud data access tutorials, I’ve added a third one (Section 3.6) which shows users how to use `intake-esm` to access data. I’ve also mentioned in the preamble of Section 3.4 (the `esgf-pyclient` and model analysis tutorial) that these other options exist and are easier to use.
-	Section 3.4 links to the ESGF web interface for searching for data.
-	Use of pandas: I agree that `xarray` is the standard (and superior) package for working with climate data. However, many new users are experienced working with tabular data but have no experience with` xarray`, which is why we started the tutorial using pandas to manipulate the station data.
-	All instances of `dask.array.apply_along_axis` have been replaced with `xr.apply_ufunc`.
-	I’ve added mention in Section 3.2 that using `chunks=’auto’` allows the user to load larger amounts of data from PAVICS via OpenDAP.
-	I’ve added instructions on accessing the AHCCD data to Section 3.1.
-	Section 3.2 links to the `xarray` and `dask` documentation on parallelization.
-	I’ve changed the PCIC data URL to the new one.
-	I tried using `xclim.ensembles.create_ensemble` to merge two datasets along the `time` dimension, but I couldn’t get it to work properly. The resulting dataset had the wrong time values for the latter half of the record. For now, I’ve mentioned that it’s possible, but I can’t add instructions on how to do it since I couldn’t get it to work.
-	All uses of `xclim.indices` have been switched to `xclim.indicators.atmos`.
-	I’ve shortened many lines which were previously too long and required scrolling.

For @brian-rose ‘s comments:
-	I’ve fixed the grammatical error and the `xclim` citation in the manuscript.
-	I’ve responded to the issues opened on the UTCDW Guidebook repository; they’ve all been closed.
-	I’ve added Section 3.6 which shows users how to access CMIP6 data using `intake-esm`.

A question for @acocac is when exactly I should merge the changes into master, create the new release, and deposit the repo in Zenodo. You said, “tag me when you complete the above steps”, but you also said that branch merge and the Zenodo deposit should take place after successful completion of the review. I assume that the reviewers need to revisit their checklists before the review is considered complete. Please advise on how to proceed.
