### 2024-07-02 #243 @aulemahal
@acocac I also confirm that I am satisfied. And I recommend this work for publication in JOSE.
