### 2024-07-03 #243 @acocac
@mikemorris12 congrats your work has been recommended for publication - The next step is to merge the branch and created a tagged release on Github. You must also archive on either https://zenodo.org/ or https://figshare.com/ and obtain a DOI (which you'll need to give to us). I can then move to the final acceptance and publication steps. I'm adding a post-review checklist which could be useful to track the progress on your publication.
