### 2024-07-04 #243 @mikemorris12
Hi @acocac, I've completed the steps of the checklist. Here is the Zenodo DOI: https://doi.org/10.5281/zenodo.12636292. Thanks for a great review experience, and thank you to @brian-rose and @aulemahal for your helpful feedback and positive reviews! I'm very pleased that this work will be published in the JOSE.
