### 2024-07-05 #243 @acocac
@mikemorris12 thanks for depositing the software in Zenodo. Before proceeding with the editor steps, can you please ensure it has the same license of the software, in this case Creative Commons Attribution Share Alike 4.0 International?

Please note you can use the [Zenodo/GitHub integration ](https://docs.github.com/en/repositories/archiving-a-github-repository/referencing-and-citing-content) for making easier the step of depositing future releases of your software.
