### 2024-07-05 #243 @mikemorris12
HI @acocac, apologies for using the wrong license on the Zenodo archive - it was previously the Creative Commons Attribution 4.0 International license, whereas the GitHub repo used the Creative Commons Attribution Share Alike 4.0 International license. They now both use the Creative Commons Attribution Share Alike 4.0 International license.
