### 2024-07-06 a910184
Update _config.yml for license in footer
