### 2024-07-06 #243 @labarba
Some inconsistencies remain in the license descriptions. The README states CC-BY-SA but it shows the CC-BY badge and link below. Also, you have code as part of the Guide, which should be licensed separately under an OSI-approved license (for example, BSD-3 or MIT for minimum restrictions).

By the way, what is the reason for the change from CC-BY to CC-BY-SA? By adding a copyleft condition, you do restrict some uses, so be sure you have good reasons for that restriction. Finally, a license change requires the explicit approval by all co-authors. In this case, it is strongly recommended that you open an issue on your repo about the license change, and get all authors to accept the change in writing there.
