### 2024-07-06 #243 @labarba
I proposed an edit to your config so that the license info appears in the footer of the Jupyter Book, rather than a standard copyright notice. See:
https://github.com/mikemorris12/UTCDW_Guidebook/pull/5
