### 2024-07-12 #243 @acocac
@mikemorris12 this is just to ask if you have any progress in considering a new license for your resource.
