### 2024-07-12 #243 @mikemorris12
Hi @acocac and @labarba, thank you for pointing out the remaining license issue, and apologies for the delay. It took a few days for the other coauthors to get back to me on this matter.

After consideration, we have decided we are obligated to use the CC-BY-SA license. This is because the Guidebook borrows materials (Specifically Chapter 3.4) from [Anderson and Smith (2021)](https://jose.theoj.org/papers/10.21105/jose.00100), which is licensed under CC-BY-SA [(see the GitLab repository)](https://gitlab.com/claut/man_ccia). This means we must use the same license.

I apologize for making what appeared to be a license change, that was merely me correcting two mistakes. One was indicating CC-BY on the Zenodo archive, and the second was using the CC-BY badge on the README page. The actual LICENSE file in the Guidebook repository was always for CC-BY-SA. Therefore I don't think we should need to open an issue about a license change, since the actual license never changed. I sincerely apologize for creating confusion about the license.

Regarding @labarba's comment that the code should be licensed separately, how should I go about doing that? Should I add a second LICENSE file to the repository? Or simply state in the README that the code materials are licensed under the MIT (for example) license? Thanks!
