### 2024-07-12 #243 @acocac
@mikemorris12 thanks for the update on this.

Re licenses, Github seems to accept multiple of them, you should follow the GitHub blog post: https://github.blog/changelog/2022-05-26-easily-discover-and-navigate-to-multiple-licenses-in-repositories/. It seems you should create a LICENSE file for the docs, and other LICENSE-CODE with the OSI-approved license for software. Please see an example in the [repo of docs.github.com](https://github.com/github/docs).

Let's wait @labarba thoughts.
