### 2024-07-19 #243 @mikemorris12
Since it's been a week and we haven't yet heard from @labarba I decided to go ahead and add the LICENSE-CODE file to the repository. (using the MIT license), following @acocac's instructions. I also added the MIT license to the Zenodo archive. Is there any further action required? Thanks!
