### 2024-07-19 #243 @acocac
> Since it's been a week and we haven't yet heard from @labarba I decided to go ahead and add the LICENSE-CODE file to the repository. (using the MIT license), following @acocac's instructions. I also added the MIT license to the Zenodo archive. Is there any further action required? Thanks!

It looks good to me. May I ask to release a new version e.g. 1.1.1 in GitHub to include the added LICENSE-CODE file? You should archive the new version in Zenodo too.
