### 2024-07-19 #243 @mikemorris12
> > Since it's been a week and we haven't yet heard from @labarba I decided to go ahead and add the LICENSE-CODE file to the repository. (using the MIT license), following @acocac's instructions. I also added the MIT license to the Zenodo archive. Is there any further action required? Thanks!
>
> It looks good to me. May I ask to release a new version e.g. 1.1.1 in GitHub to include the added LICENSE-CODE file? You should archive the new version in Zenodo too.

Thanks! I've created a [new release](https://github.com/mikemorris12/UTCDW_Guidebook/releases/tag/v1.1.1) and a new Zenodo version (https://doi.org/10.5281/zenodo.12785645).
