### 2024-07-19 #243 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
