### 2024-07-25 #243 @mikemorris12
Hi @acocac, sorry for the delay. I had trouble checking off the boxes in @labarba's message so I've copied the author checklist here:

- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSE paper. [(version 1.1.1)](https://github.com/mikemorris12/UTCDW_Guidebook/releases/tag/v1.1.1)
- [x] Archive the release on Zenodo/figshare/etc and post the DOI here. [(Zenodo Archive link)](https://doi.org/10.5281/zenodo.12785645)

I've left the last item unchecked since the license listed in the manuscript does not match the license in the Zenodo Archive or the GitHub repository. The manuscript lists CC-BY, not CC-BY-SA. Is the license for the manuscript different from the license for the software/guidebook? If so, then consider the last item checked off, since the both GitHub and Zenodo list CC-BY-SA as the license.
