### 2024-07-31 #243 @acocac
@labarba can you reply to the latest author's comment?

> I've left the last item unchecked since the license listed in the manuscript does not match the license in the Zenodo Archive or the GitHub repository. The manuscript lists CC-BY, not CC-BY-SA. Is the license for the manuscript different from the license for the software/guidebook? If so, then consider the last item checked off, since the both GitHub and Zenodo list CC-BY-SA as the license.
