### 2024-07-31 #243 @labarba
The license of the paper is CC-BY per JOSE policy, while the license of the software/content is up to the authors (but must be OSI-approved for software and a standard public license for content).
