### 2024-08-01 #243 @editorialbot
:wave: @openjournals/jose-eics, this paper is ready to be accepted and published.
