### 2024-08-16 #243 @mikemorris12
Hi @acocac and @labarba, it's been about 2 weeks since the last correspondence. It looks like the editorial bot is waiting for a command. Are there any other steps for me to take for the paper acceptance to be finalized? Thanks in advance!
